Use a better spots array -> transfer computational load (not that much either tbh too) to the split function Thus we reduce accesses in counting from 4 (2 in sa and 2 in spots) to just 2 (1 in sa and 1 in spots) per match. Furthermore, the splits function also presents the opportunity to combine with the parsing subroutine and although parsing an spliting are both O(n), we can hopefully reduce number of operation from O(2n) to O(n). With this optimization we drop to ~4s flat on book1 on my machine. (no multithreading yet)

+++ src/match_finder.rs
@@ -121,8 +121,8 @@ impl Match {
         self.sa_count_len >> 8
     }
 
-    pub fn get_len(&self) -> u32 {
-        self.sa_count_len & 0xff
+    pub fn get_len(&self) -> i32 {
+        (self.sa_count_len & 0xff) as i32
     }
 }
